requirements.txt
streamlit>=1.40.0
streamlit-aggrid==0.3.4.post3
pandas
gspread
google-auth
google-genai
python-dotenv
openpyxl
xlrd
